refactor(dogstatsd): gate replay behind linux flags and make replaysend more generic
