+++ bin/agent-data-plane/src/cli/utils.rs
@@ -1,7 +1,12 @@
 use futures::TryFutureExt as _;
 use http::{header::CONTENT_TYPE, uri::PathAndQuery, Request, Response, StatusCode, Uri};
-use http_body_util::{BodyExt as _, Full};
-use hyper::body::{Bytes, Incoming};
+use http_body_util::BodyExt as _;
+#[cfg(target_os = "linux")]
+use http_body_util::Full;
+#[cfg(target_os = "linux")]
+use hyper::body::Bytes;
+use hyper::body::Incoming;
+#[cfg(target_os = "linux")]
 use prost::Message as _;
 use saluki_config::GenericConfiguration;
 use saluki_error::{generic_error, ErrorContext as _, GenericError};
@@ -29,6 +34,7 @@ struct DogStatsDCaptureResponseBody {
     path: String,
 }
 
+#[cfg(target_os = "linux")]
 #[derive(Deserialize)]
 struct DogStatsDReplaySessionResponseBody {
     session_id: String,
@@ -228,6 +234,7 @@ impl DataPlaneAPIClient {
     ///
     /// If the request fails, if ADP rejects the session, or if the response body can't be decoded, an error is
     /// returned.
+    #[cfg(target_os = "linux")]
     pub async fn dogstatsd_replay_start_session(
         &mut self, state: Option<&datadog_protos::agent::TaggerState>,
     ) -> Result<String, GenericError> {
@@ -254,6 +261,7 @@ impl DataPlaneAPIClient {
     /// # Errors
     ///
     /// If the request fails, or if ADP rejects the session release, an error is returned.
+    #[cfg(target_os = "linux")]
     pub async fn dogstatsd_replay_finish_session(&mut self, session_id: &str) -> Result<(), GenericError> {
         let uri = self.build_uri(&format!("/dogstatsd/replay/session/{session_id}"), None);
         let req = Request::delete(uri)
@@ -357,6 +365,7 @@ fn body_when_capture_success(resp: Response<String>) -> Result<String, GenericEr
     }
 }
 
+#[cfg(target_os = "linux")]
 fn body_when_replay_session_success(resp: Response<String>) -> Result<String, GenericError> {
     match resp.status() {
         status if status.is_success() => Ok(resp.into_body()),
@@ -395,6 +404,7 @@ fn empty_when_success(resp: Response<String>) -> Result<(), GenericError> {
     }
 }
 
+#[cfg(target_os = "linux")]
 fn empty_when_replay_session_success(resp: Response<String>) -> Result<(), GenericError> {
     match resp.status() {
         status if status.is_success() => Ok(()),
@@ -411,7 +421,9 @@ fn empty_when_replay_session_success(resp: Response<String>) -> Result<(), Gener
 mod tests {
     use http::{Response, StatusCode};
 
-    use super::{body_when_capture_success, body_when_replay_session_success, empty_when_replay_session_success};
+    use super::body_when_capture_success;
+    #[cfg(target_os = "linux")]
+    use super::{body_when_replay_session_success, empty_when_replay_session_success};
 
     #[test]
     fn dogstatsd_capture_failed_precondition_surfaces_server_message() {
@@ -448,6 +460,7 @@ mod tests {
         assert!(message.contains("route not found"));
     }
 
+    #[cfg(target_os = "linux")]
     #[test]
     fn dogstatsd_replay_session_conflict_surfaces_server_message() {
         let response = Response::builder()
@@ -459,6 +472,7 @@ mod tests {
         assert_eq!(error.to_string(), "DogStatsD replay already in progress.");
     }
 
+    #[cfg(target_os = "linux")]
     #[test]
     fn dogstatsd_replay_finish_conflict_surfaces_server_message() {
         let response = Response::builder()
